Add WLEDSoftwareUpdateRequest
