+++ wled-native/Model/DeviceApi/Request/WLEDChangeStateRequest.swift
@@ -3,11 +3,6 @@ import Foundation
 import CoreData
 
 struct WLEDChangeStateRequest: WLEDRequest {
-    let context: NSManagedObjectContext
     let state: WLEDStateChange
-    
-    init(state: WLEDStateChange, context: NSManagedObjectContext) {
-        self.state = state
-        self.context = context
-    }
+    let context: NSManagedObjectContext
 }
