+++ wled-native/Model/DeviceApi/Request/WLEDRefreshRequest.swift
@@ -4,8 +4,4 @@ import CoreData
 
 struct WLEDRefreshRequest: WLEDRequest {
     let context: NSManagedObjectContext
-    
-    init(context: NSManagedObjectContext) {
-        self.context = context
-    }
 }
