+++ wled-native/Model/DeviceApi/Request/WLEDSoftwareUpdateRequest.swift
@@ -1,0 +1,10 @@
+
+import Foundation
+import CoreData
+
+struct WLEDSoftwareUpdateRequest: WLEDRequest {
+    let context: NSManagedObjectContext
+    let binaryFile: URL
+    let onCompletion: () -> ()
+    let onFailure: () -> ()
+}
